[MIG][12.0] pos_payment_terminal
Update pos_payment_terminal/static/src/js/devices.js


[REF] rename field

[UPD] Update pos_payment_terminal.pot

[UPD] README.rst

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: pos-12.0/pos-12.0-pos_payment_terminal
Translate-URL: https://translation.odoo-community.org/projects/pos-12-0/pos-12-0-pos_payment_terminal/

[UPD] Update pos_payment_terminal.pot

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: pos-12.0/pos-12.0-pos_payment_terminal
Translate-URL: https://translation.odoo-community.org/projects/pos-12-0/pos-12-0-pos_payment_terminal/

Translated using Weblate (Spanish)

Currently translated at 100.0% (10 of 10 strings)

Translation: pos-12.0/pos-12.0-pos_payment_terminal
Translate-URL: https://translation.odoo-community.org/projects/pos-12-0/pos-12-0-pos_payment_terminal/es/

Update translation files

Updated by "Update PO files to match POT (msgmerge)" hook in Weblate.

Translation: pos-12.0/pos-12.0-pos_payment_terminal
Translate-URL: https://translation.odoo-community.org/projects/pos-12-0/pos-12-0-pos_payment_terminal/
