+++ pos_payment_terminal/models/pos_order.py
@@ -1,9 +1,10 @@
 # © 2018 ACSONE SA/NV
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
+
 from collections import defaultdict
 import logging
 
-from odoo import models, api
+from odoo import api, models
 
 _logger = logging.getLogger(__name__)
 
@@ -20,7 +21,7 @@ class PosOrder(models.Model):
         currency_digits = pos_session.currency_id.decimal_places
         card_journals = self.env['account.journal'].search([
             ('id', 'in', [p[2]['journal_id'] for p in payments]),
-            ('payment_mode', '!=', False),
+            ('pos_terminal_payment_mode', '!=', False),
         ])
         card_payments = [record[2] for record in payments
                          if record[2]['journal_id'] in card_journals.ids]
