+++ pos_payment_terminal/static/src/xml/pos_payment_terminal.xml
@@ -2,7 +2,7 @@
 <templates id="template" xml:space="preserve">
     <t t-extend="PaymentScreen-Paymentlines" >
         <t t-jquery=".col-name" t-operation="append">
-            <t t-if="line.cashregister.journal.payment_mode and widget.pos.config.iface_payment_terminal">
+            <t t-if="line.cashregister.journal.pos_terminal_payment_mode and widget.pos.config.iface_payment_terminal">
                    <button class="payment-terminal-transaction-start"  t-att-data-cid='line.cid'>Start transaction</button>
             </t>
         </t>
