+++ pos_payment_terminal/views/account_journal.xml
@@ -1,14 +1,15 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <record id="view_account_journal_pos_user_form" model="ir.ui.view">
-        <field name="name">pos.payment.terminal.journal.form</field>
+    <record id="view_account_journal_form" model="ir.ui.view">
         <field name="model">account.journal</field>
         <field name="inherit_id" ref="point_of_sale.view_account_journal_pos_user_form"/>
         <field name="arch" type="xml">
-            <field name="amount_authorized_diff" position="after">
-                <field name="payment_mode"/>
-            </field>
+            <group name="amount_authorized" position="after">
+                <group name="payment_mode_group" attrs="{'invisible': [('type', '!=', 'bank')]}">
+                    <field name="pos_terminal_payment_mode"/>
+                </group>
+            </group>
         </field>
     </record>
